Moved the definition of the find_free_port Python function from set_connectivity_service_port.py to utils.py so that it is available for use in other scripts.
