+++ python/daqconf/set_connectivity_service_port.py
@@ -1,8 +1,8 @@
 import conffwk
 import confmodel
+from daqconf.utils import find_free_port
 
 import re
-import socket
 
 def set_connectivity_service_port(oksfile, session_name, connsvc_port=0):
     """Script to set the value of the Connectivity Service port in the specified Session of the specified
@@ -24,14 +24,6 @@ def set_connectivity_service_port(oksfile, session_name, connsvc_port=0):
     dal = conffwk.dal.module("dal", schemafiles)
 
     if connsvc_port == 0:
-        def find_free_port():
-            with socket.socket() as s:
-                s.bind(("", 0))
-                s.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
-                port = s.getsockname()[1]
-                s.close()
-                return port
-
         new_port = find_free_port()
     else:
         new_port = connsvc_port
